backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
PyPDF2==3.0.1
Werkzeug==2.3.7
supabase>=2.24.0
bcrypt==4.1.2
PyJWT>=2.10.1
python-dotenv==1.0.1
websockets>=15.0.1